Allow transformer to skip entries by returning `nil` or `false`
I needed to skip some elements in transform phase, so I added this feature.
When implementing it here I just noticed that it was already possible to do that by returning and empty array in transformer, but I decided to still add the feature here as returning `nil` is more obvious for the API consumer.